kafka-producer/requirements.txt
kafka-python==2.0.2
aiokafka==0.7.2
Faker==15.1.3